Disable MSVC fastlink
Item icons - checked, checked hovered, and checked pressed states
MOBILE-252: Fix content displaying under Android software system buttons.
MOBILE-273: Change warning appearance for cloud screens.
MOBILE-274: Fix content displayed under android title bar on some devices.
MOBILE-277: Cloud account offline behavior.
MOBILE: Display field warning in custom session screen instead of warning panel.
MOBILE: Fix invalid credentials warning after cloud log out.
Notification widgets reacted to doubleclicks instead of clicks.
QFileDialog table header fix.
QnInputField could paint its input red while keeping focus.
Review fix.
VMS-2163: Make mobile client use QnConnectionValidator.
VMS-2355 - wrong server is dispalyed in connected servers section of the About window
VMS-2436 - 'Other systems' icons
VMS-2442: Add border under panels headers
VMS-2531 - mic button for 2-way-audio on item at Layout for unauthorised devices isn't displayed without item reopening
VMS-2581 - Hide 'buggy' icon for viewers and all tree icons in selection lists
VMS-2583 - Toolbar icons at layout items doesn't display its state
VMS-2640: Add line inder the tree header
VMS-2742: Implement tooltips style.
VMS-2814: Visually unpleasant behavior of text area placeholder in 'Text overlay' rule settings.
VMS-2856 - Fix recorder nodes display in tree
VMS-3099: Fix camera addition dialog layout
VMS-3135 - hi-dpi: window is jumping and flashing when drag between monitors (same screens scale)
VMS-3139: hi-dpi: elements are not accessible
VMS-3309 - Fix behaviour when confirming changes and disconnect
VMS-3335: Fix update progress bar (System Administration -> Updates)
VMS-3355 - Fix 'email is not set' notifications
VMS-3418 - 'Hide panel' button jumps on right panel
VMS-3474: Improve manual update process.
VMS-3480: Update Tab redesign
VMS-3571 - Prepare for Temporary credentials: fix cloud account panels
VMS-3589 - NX client crash
VMS-3670: Fix hidpi other screen bugs : fix rotation instrument click
VMS-3670: Fix hidpi other screen bugs : invisible item when moving
VMS-3678: HiDpi
VMS-3678: HiDpi. build fix
VMS-3678: HiDpi. remove tmp
VMS-3678: Wrong menu positioning on hidpi-not-first screen
VMS-3683 - nx client crash
VMS-3689 - custom user can open any shared layout, even that he hasn't access to
VMS-3708 - HiDPI: non-clickable menus in dialogs.
VMS-3740 - user list for business rules action target users mustn't be expandable
VMS-3749: Change field order in 'Change password' dialog
VMS-3751: Disable content on fisheye tab when dewarping is turned off
VMS: Compilation fix for g++.
fadeWidget service function.
tmp

<==== Mercurial data ====>
Branch: dev_3.0.0

+++ client/libclient/src/ui/dialogs/backup_schedule_dialog.cpp
@@ -36,11 +36,11 @@ namespace {
         return mBits * bitsPerMegabit / bitsPerByte;
     }
 
-    int dayToIndex(Qt::DayOfWeek day) {
-        return static_cast<int>(day) - 1;
+    size_t dayToIndex(Qt::DayOfWeek day) {
+        return static_cast<size_t>(day) - 1;
     }
 
-    Qt::DayOfWeek indexToDay(int index) {
+    Qt::DayOfWeek indexToDay(size_t index) {
         return static_cast<Qt::DayOfWeek>(index + 1);
     }
 }
